README.h ========= Cleaning up information
ViewController.swift
===============
Cleaning up comments.  Renamed a couple functions.  Reorganized some code.

moviePlayerWillExitFullscreenNotification
Simplifying code.

Changed some code to reflect that we are more interested in determining if a source video has a 90 or 270 degree rotation
than we are in whether a source video is Portrait or Landscape (i.e. when deciding whether to swap width/height during export).