ViewController.swift +++++++++++++++++++ Added videoAsset as a stored property.
playMovie
========
Fixed a bug wherein we need to remove previous player from super view if it exists.

processCameraMovie
=================
Fixed a couple bugs.

videoURL needs to reference the stored property so that it is available to the export method.

videoAsset needs to be a stored property to avoid an issue with completion handler of loadValuesAsynchronouslyForKeys not being called consistently.

processLibraryMovie_0
===================
videoAsset needs to be a stored property here also for same reason as in processCameraMovie

processLibraryMovie_1
==================
Now referencing self.videoAsset